+++ opentelemetry-proto/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Implement tonic metrics proto transformations (#1184)
 - Move proto for zPage to tonic [#1214](https://github.com/open-telemetry/opentelemetry-rust/pull/1214)
+- Support exponential histograms (#1267)
 
 ### Changed